+++ cerberus-server/src/main/java/com/cerberus/server/message/RFIDAuthResponseMessage.java
@@ -9,7 +9,7 @@ public class RFIDAuthResponseMessage extends Message {
 	private final boolean isAuthorized;
 
 	@JsonCreator
-	public RFIDAuthResponseMessage(@JsonProperty("socketId") int socketId, @JsonProperty("timestamp") long timestamp,
+	public RFIDAuthResponseMessage(@JsonProperty("socketId") long socketId, @JsonProperty("timestamp") long timestamp,
 			@JsonProperty("rfidNumber") int rfidNumber, @JsonProperty("authorized") boolean isAuthorized) {
 		super(MessageType.RFID_AUTH_RES, socketId, timestamp);
 		this.rfidNumber = rfidNumber;
