+++ cerberus-server/src/test/java/com/cerberus/server/message/SwitchOperatingModeMessageTest.java
@@ -1,10 +1,9 @@
 package com.cerberus.server.message;
 
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -13,12 +12,12 @@ import com.cerberus.server.logic.constants.SocketOperatingMode;
 
 public class SwitchOperatingModeMessageTest {
 
-	private int socketId;
+	private long socketId;
 	private long timestamp;
 
 	@Before
 	public void setup() {
-		socketId = 12345;
+		socketId = 12345L;
 		timestamp = System.currentTimeMillis();
 	}
